+++ src/classes/event/OnlineEvent.java
@@ -8,7 +8,7 @@ public class OnlineEvent extends Event {
     //Constant name convention
     private final Location location;
 
-    public OnlineEvent(int ID, String title, LocalDateTime start, long length, Unit unit, String[] participants, Location location) {
+    public OnlineEvent(int ID, String title, LocalDateTime start, int length, Unit unit, String[] participants, Location location) {
         super(ID, title, start, length, unit, participants);
         this.location = location;
     }
