+++ src/classes/event/OnsiteEvent.java
@@ -9,7 +9,7 @@ public class OnsiteEvent extends Event {
     private Location location;
 
 
-    public OnsiteEvent(int ID, String title, LocalDateTime start, long length, Unit unit, String[] participants, Location location) {
+    public OnsiteEvent(int ID, String title, LocalDateTime start, int length, Unit unit, String[] participants, Location location) {
         super(ID, title, start, length, unit, participants);
         this.location = location;
     }
